@@ -2,7 +2,7 @@ CMAKE_MINIMUM_REQUIRED(VERSION 3.18)
 PROJECT("RogEng")
 OPTION(Global "Global" OFF)
 SET(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-SET(CMAKE_CXX_STANDARD 17)
+SET(CMAKE_CXX_STANDARD 20)
 ADD_LIBRARY(RogEng "${SRC}")
 SET(SRC
     src/RogEng/RE.cpp
